Update patch set 18
Patch Set 18: Code-Review+2

I think this is a good enough check-point, especially with the understanding that this whole package's API is still experimental and subject to change (with that new disclaimer at the top of the package doc).

Thanks again for all your effort on this.

Patch-set: 18
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2