Update patch set 1
Patch Set 1:

A general question first:

Why have a renew loop per domain? Maybe we were optimizing prematurely (although I think it was just the first implementation that occurred to us), but my other toy library had a single goroutine that was blocked on a select with two channels (a timeout channel set to the minimum renewal time of all certs in the map, and a buffered channel that was notified everytime a cert was added to the certificate map).

Patch-set: 1
Label: Code-Review=0