Update patch set 2
Patch Set 2:

goroutines are cheap, event if you ran 1000s of domains, and I thought it would be easier to understand one renew loop per domain implementation, seemed more clear.

But, of course, it's subjective and I could be just attached to this implementation that has been running in my head for the past 2 days.

(the new patchset is only to remove a leftover typo)

Patch-set: 2