Update patch set 6
Patch Set 6:

I analyzed the output of Go 1.7 compared with the manual assembly. There is basically no difference if you exclude the fact that the generated code has some bound checks (because of an array of constants) while the manual assembly is fully unrolled so constants are inlined. It looks like we're very close at producing perfect code with the compiler. Maybe fixing the prove pass to handle this situation and drop the bound checks is enough to achieve the same performance.

Patch-set: 6