Update patch set 2
Patch Set 2:

x/crypto has a fairly low bar: packages should plausibly be useful to a fair number of people who want to interoperate with existing code. (Or else be something that we wish to encourage the use of in future.)

For HC and Serpent, however, I'm not sure that bar is met. I'm not aware of any common protocols that use them and suspect this code might be best in a personal repo, I'm afraid.

Patch-set: 2
Label: Code-Review=0