Update patch set 1
Patch Set 1:

This is cool work, but I was waiting to see whether it ended up being used for anything. I would like to see CONIKS in use, but I don't think that's currently the case. In light of that, it seemed that this might be best suited for a personal repo.

(If there are deployments of CONINKs, or other cool uses of vrfs, that I'm missing then please let me know.)

Patch-set: 1