chacha20poly1305: test both SSSE3 and SSE4.1 support on amd64.
The non-AVX code in chacha20poly1305_amd64.s uses PINSRB, which was
introduced in SSE4.1. This change adds a test for SSE4.1 before we try
and use the assembly code.

It might be that all CPUs that support SSE4.1 must also support SSSE3,
but the nearest statement that I can find from Intel is that “SSE4.1 and
SSE4.2 are extensions of SSE, SSE2, SSE3, and SSSE3.”[1], which I don't
think is strong enough to assume that.

[1] https://software.intel.com/sites/default/files/m/8/b/8/D9156103.pdf

Fixes #17464

Change-Id: If5448ede07ceac24d2b9a88d3de54ded6df6a63b