Update patch set 4
Patch Set 4:

> Uploaded patch set 4.

I took a stab at implementing a `Available` check but found no way to actually test it and pulled it out.

As far as I can tell the only way this could happen would be if one of the SHA packages had a build tag that meant on certain systems one of the hashes wouldn't be registered but a cursory search didn't reveal any.

Patch-set: 4