Update patch set 8
Patch Set 8:

> > Patch Set 8:
 > >
 > > 2-3% improvement doesn't seem worthwhile enough to pay the cost
 > of adding 211 lines of code and new +build tag rules.
 > 
 > Perhaps we should set some rule regarding assembly optimizations.
 > IMHO, less than 10% gain is never worth the trouble, esp. if
 > the new code requires more recent instruction set extensions.

In general, yes - it's quite a lot extra complexity for 3% performance. 
I just updated the implementation to use BMI2 (like chacha20poly1305) - poly1305 is used in NaCl (with Salsa20) and in combination with AES.
On systems using AES+poly1305 (with NI), poly1305 is the limiting component.
But I agree, it's a trad-off between complexity and performance - and 3% is not much.

Patch-set: 8