ssh: rewrite (re)keying logic.
Use channels and a dedicated write loop for managing the rekeying
process.  This lets us collect packets to be written while a key
exchange is in progress.

Previously, the read loop ran the key exchange, and writers would
block if a key exchange was going on. If a reader that would write
back a packet while processing a read packet, it could block, stopping
the read loop, thus causing a deadlock.

Unfortunately, such coupled read/writes are inherent with handling
requests that want a response (eg. keepalive, opening/closing
channels etc.)

The new scheme means that writePacket always returns immediately, so
there is no backpressure on fast writers. Since SSH has flow control
for normal streams, this is only an issue for processes that send OOB
requests messages as fast as they can.

Fixes #18439.

Change-Id: I7c14ff4991fa3100a5d36025125d0cf1119c471d