x/ssh: rewrite (re)keying logic.
Use channels and a dedicated write loop for managing the rekeying
process.  This lets us collect packets to be written while a key
exchange is in progress.

Previously, the read loop ran the key exchange, and writers would
block if a key exchange was going on. If a reader that would write
back a packet while processing a read packet, it could block, stopping
the read loop, thus causing a deadlock.

This has the following implications:

* A minor performance degradation: since handshakeTransport now queues
  writePacket calls, writePacket must always be supplied with a fresh
  buffer. We have to let go of the packetPool in channel.

* Since writePacket now returns immediately, there is no backpressure
  on fast writers. This is only an issue for processes that send OOB
  requests messages as fast as they can (OOB messages are not subject
  to flow control).

TODO - benchmarks

Fixes #18439.

Change-Id: I7c14ff4991fa3100a5d36025125d0cf1119c471d