x/ssh: rewrite (re)keying logic.
Use channels and a dedicated write loop for managing the rekeying
process.  This lets us collect packets to be written while a key
exchange is in progress.

Previously, the read loop ran the key exchange, and writers would
block if a key exchange was going on. If a reader that would write
back a packet while processing a read packet, it could block, stopping
the read loop, thus causing a deadlock.

Fixes #18439.

TODO:
 * fix tests
 * check that all error conditions lead to exit.

Change-Id: I7c14ff4991fa3100a5d36025125d0cf1119c471d