x/crypto/ssh: support MaxAuthTries on ServerConfig
MaxAuthTries specifies the maximum number of authentication attempts
permitted per connection. If set to a number less than or equal to 0,
the server will allow unlimited authentication attempts. This is to
support existing servers that are not currently setting this property.
This configuration property mirrors a similar property in sshd_config
and prevents bad actors from continuously trying authentication. On
exceeding maximum authentication attempts, the server will send a
disconnect message to the client.

Change-Id: Ic77d2c29ee2fd2ae5c764becf7df91d29d03131b