Update patch set 1
Patch Set 1: Code-Review+1

With this code on s390x I get about a 10% slowdown on the 64 byte benchmarks and a 5% speedup on the 1K benchmarks (vs. the original reference). As you noted we'll probably add an assembly version for s390x at some point, but in the meantime this seems like a reasonable tradeoff, especially given how much simpler the new code is.

Patch-set: 1
Label: Code-Review=+1