Update patch set 1
Patch Set 1:

> (are there recommendations for RC4 rekeying intervals?)

The only RFC (expressly considered RC4 for SSH I know) is RFC 4345, but there are no recommendations about rekeying.

RC4 (as a stream-cipher) is not affected by birthday attacks (like block ciphers) - so I suspect the recommendations of RFC 4344 should only apply to block ciphers. 
But RC4 is broken and shouldn't used anymore, so it's probably reasonable to treat RC4 the same way (or not worth to treat it different).
Or missed I something, Adam?

Patch-set: 1