Update patch set 1
Patch Set 1:

Thanks all. I asked Martin to draft this up in order to potentially replicate the good experience that we've had with a similar API in BoringSSL. We'll likely do a few rounds of improvements before it's really suitable for public review.

(For background: the motivation is that crypto/tls has a lot of hand-crafted marshal and unmarshal code which isn't terribly performance critical, but which is a bit of a burden. Also, my feeling is that the design of encoding/asn1 was a mistake. Thus we might want a replacement for the std lib's internal use.)

Patch-set: 1
Label: Code-Review=0