Update patch set 5
Patch Set 5: Code-Review+1

> Patch Set 5:
> 
> is this necessary? the callback could be a method on a struct that holds a counter.  You could then have different counts (eg. a max try on a public-key doesn't make much sense).

Any ETA on when this is getting reviewed? We're suffering from this on our environments and the only way to get around it is by restarting the ssh server processes. And even then it's only a matter of time before it happens again.

This implementation is the same strategy OpenSSH uses so it seems like a legit solution.

Patch-set: 5
Reviewer: Gerrit User 17530 <17530@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1