Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 5:
> 
> (1 comment)
> 
> If we add this, I think we should also provide a  > 0 default. Otherwise, you still have to understand the need for MaxAuthTries in order to get the security benefits for it.

I agree. I updated the change to default MaxAuthTries to 6 to mirror OpenSSH. This only takes effect when the MaxAuthTries on the ServerConfig is set to 0. I left the behavior for when MaxAuthTries is set to less than 0 as a manner for users to turn this functionality off.

Patch-set: 6
Reviewer: Gerrit User 16855 <16855@62eb7196-b449-3ce5-99f1-c037f21e1705>