ssh: improve client public key authentication
Previously, the public key authentication for clients would send an
enquiry to the remote for every key specified and build up a list of
valid keys before attempting to authenticate with the server. This
change updates this logic to short circuit the process once a valid key
is found. If authentication fails after finding a valid key, the client
will repeat this process with the remaining keys.

Change-Id: I8ea42caf40c0864752218c3f6934e86b12f5b81a