ssh: improve client public key authentication
Previously, the public key authentication for clients would send an
enquiry to the remote for every key specified and build up a list of
valid keys before attempting to authenticate with the server.

This change updates client public key authententication logic to short
circuit the process once a valid key is found. If authentication fails
after finding a valid key, the client will repeat this process with the
remaining keys. This implementation will result in two fewer packets
from being exchanged by the client and server for every key that follows
a valid and successfully authenticated key. In the case in which all
keys fail authentication, the number of packets exchanged by the client
and server is unaffected.

Also for daemons that are implemented by OpenSSH, an enquiry into the
validity of a key without an authentication attempt will still be
recorded as an authentication attempt by the daemon. This will result in
OpenSSH rejecting any public key authentication attempt from a client in
which the client is attempting to authenticate with a list of keys of
length longer than the MaxAuthTries configuration of the OpenSSH daemon.
This change will allow clients to succeed authentication with OpenSSH
daemons if the successful key is positioned at an index smaller than the
MaxAuthTries configuration.

Change-Id: I8ea42caf40c0864752218c3f6934e86b12f5b81a