Update patch set 4
Patch Set 4: Run-TryBot+1

(2 comments)

I tried to rewrite the message, but looks like I have no permissions to edit messages. How about this:



ssh: improve client public key authentication

Previously, the public key authentication for clients would send an
enquiry to the remote for every key specified before attempting to
authenticate with the server.

Now, we immediately try to authenticate once a valid key is found. 
This results in exchanging fewer packets if the valid key is near the 
top of the list. If all keys fail, then the number of packets exchanged 
by the client and server is unaffected.

For OpenSSH daemon, an enquiry into the validity of a key is recorded
as an authentication attempt, so any clients with more than MaxAuthTries
public keys would not be able to authenticate before. This change will 
allow  clients to succeed authentication if the successful key is at the 
start of the list of keys.

Patch-set: 4
Reviewer: Gerrit User 5893 <5893@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1