Update patch set 1
Patch Set 1:

I'm not sure it's a good idea for golang.org/x/crypto to depend on internal details of the runtime package.  The runtime package may change those details--for example, CPU detection may all move to a runtime/internal/cpuid package--and then we will need build tags and the like to keep this code working.  In the absence of a defined API, better to avoid internal details.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>