Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I'm not sure it's a good idea for golang.org/x/crypto to depend on internal  details of the runtime package.  The runtime package may change those details--for example, CPU detection may all move to a runtime/internal/cpuid package--and then we will need build tags and the like to keep this code working.  In the absence of a defined API, better to avoid internal details.
> 
> Okay, thanks for the hint. I agree that this should wait until there's a defined API - I just thought that the API is already stable because for example https://go-review.googlesource.com/c/39953/ does the same in the standard library.

Thats is depending on the runtime implementation detail (there is no api) and that the std lib release is coupled with the gc compiler release. So whoever changes the runtime has the work to fix all of std lib at the same time currently. As said in the other CL there is work underway to have an internal /cpu api in the std lib but is not yet committed. However that would not provide a public api to outside the std lib either.

Patch-set: 1