Update patch set 1
Patch Set 1:

> (1 comment)
Thanks for the hint. As this is go1.7 and higher we can indeed use the asm supported version.
 
 > Is there a way to use this in other parts of crypto (e. g.
 > blake2b)?

I am working on an internal package just for go std lib in https://golang.org/cl/41476/.
a public package seems discouraged see: https://github.com/golang/go/issues/19739

However not sure what the opionons are and if its possible to have a cloned internal/cpu package (once commited in std/lib) under crypto that is not public outside of crypto.
( But then if we vendor that in we end up with two cpu feature detection packages in the std/lib unless we adjust the imports ).

I would still like to remove the dependency here in this CL independently of any later refactoring as this would allow to vendor in the new version and complete the work of feature detection seperation from the runtime in the std go lib by go1.9.

Patch-set: 1
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>