Update patch set 3
Patch Set 3:

Sorry, I'm in a summit this week and a conference next week. I'll try to get to this, but no promises. This does look like it warrants a thorough review, so I want to give it more than a glance.

Patch-set: 3
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>