crypto/ssh: only check principals against hostname
prior to this, if a host running on a non-standard port wanted to use an
ssh certificate, crypto/ssh would require that the ValidPrincipals
contained the non-standard portnumber. eg.

  ValidPrincipals:
   a.b.c
   a.b.c:1234

Now a hostcert is good for any port on a given principal. This is closer
to what openssh does.

Note: for a client to accept this cert, if they're using the golang ssh
knownhosts library, their known_hosts file still needs to contain the
non-standard port, eg

  @cert-authority a.b.c,[a.b.c]:1234 ssh-ed25519 <key>

fixes issue 20273

Change-Id: I305cbbb82296b99259285e03399c9498494bc700