ssh: bypass nil publicKeyCallback functions while attempting to authenticate with crypto/ssh public keys
It is possible to define a publicKeyCallback function that returns an empty slice of `ssh.Signer`s, creating
a nil pointer dereference. This patch checks for nil functions before attempting to call the function. This
patch also leaves a comment that explains why no error is returned in this case.

Fixes #20441.

Change-Id: I19617b88e7667f7728b9a53ef9b794c99d0ae880