Update patch set 1
Patch Set 1:

I don't think this is the right place for this fix, whatever is letting a nil callback should have done this verification earlier. 

With that said, it's not clear to me at the moment want the linked bug is, I'd like to see some discussion (and reproduction) of the bug happen before jumping to coding.

Patch-set: 1
CC: Gerrit User 5150 <5150@62eb7196-b449-3ce5-99f1-c037f21e1705>