Update patch set 3
Patch Set 3: Code-Review+1

Patch set 3 looks ideal to me. As discussed elsewhere, I agree that 160- and 224-bit support is a bit needless, and that 128-bit is only useful as a MAC, so the restrictions you've given it are reasonable.

I'd wait for Mathias to chime in with his approval, as he'll be able to test whether or not it's useful and correct immediately.

Patch-set: 3
Reviewer: Gerrit User 20556 <20556@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1