Update patch set 2
Patch Set 2:

(1 comment)

I wonder if adding an example is the right approach. Maybe it is better to make the doc string of Marshal more clear that it handles the spec in https://www.ietf.org/rfc/rfc4251.txt section 5. It is provided for compatiblity with extensions of the SSH protocol that use the same data layout. 

People that can define their own protocol extension should just use JSON or something.

Patch-set: 2