Update patch set 2
Patch Set 2:

Thanks.

One other thing occurred to me. Should Verify take a slice rather than an array pointer for the authenticator? I have tended to use array pointers in several places because they allow the type-system to do more, but I had always assumed that Go would grow the ability to cast from a slice to an array pointer (with a run-time check, as usual). However, it doesn't appear that's going to happen so I do it less often now.

To put it another way: I'm inclined to change Verify to take a slice for the authenticator before landing. Any objections?

(I think it's fine for Sum to return an array pointer because that's little different than returning a slice for the code using this package.)

Patch-set: 2
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>