Update patch set 2
Patch Set 2:

> > I wish Go forced some kind of quotation around arguments (or
 > globals) that are referenced inside a function/method comment.
 > >
 > > Like `a' (Emacs Lisp style) or @p a (doxygen style), concrete
 > syntax does not matter, any one would work for me. But it is too
 > late, I suppose.
 > 
 > Yeah, it's very helpful. Is it possible to plan to add the
 > quotation as a new feature?

I think it is worthwhile to check issue tracker for existing issue first. If there is none, you can create a new one, maybe even prepend "proposal: " to it's title, and describe the problems with current approach and benefits of the alternative. This way, you will, most likely, get an answer to your question.

I will be glad if you CC me there.

Patch-set: 2