Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> this needs client support too (The ClientConfig should have some kind of callback), and test that runs the client against the server, checking that the banner gets received.

Hi,

I needed the client support for SSH banner for one of my projects.
So I implemented it with the associated test and the conflict of this CL resolved, yet I can't submit a CL to gerrit and keep joshua authorship with a specific commit.
How should I proceed ?

Thanks

Patch-set: 1
CC: Gerrit User 5893 <5893@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 23797 <23797@62eb7196-b449-3ce5-99f1-c037f21e1705>