ssh: add DialThroughJumphosts
proxy.Dial will connect to the last host in the addresses
slice by jumping through each of the preceding hosts.

  proxy.Dial("tcp", []string{B, A}, cfg)

is the equivalent of something like

  ssh A -o ProxyCommand=ssh B -W A:22

but given that this takes a slice, it's easy to do something more
exotic like

  proxy.Dial("tcp", []string{E, D, C, B, A}, cfg)

to connect to A through hosts E, D, C & B, in that order.

Change-Id: Ifcd25e7493f70ca04532bab0e57b9935f4eca261