Update patch set 1
Patch Set 1:

it looks useful, but I'd rather see this in an external library and move into the core package when it gets popular. Since it is a client-side construct, you don't really have to be in the ssh package, do you?

Patch-set: 1
Reviewer: Gerrit User 5893 <5893@62eb7196-b449-3ce5-99f1-c037f21e1705>