Update patch set 4
Patch Set 4:

(6 comments)

Perhaps an int64 would be sufficient. What drove this was an RFC that required 4294967296 iterations. This would work fine on a 64bit system with int but would fail on a 32bit system. I was therefore looking to work around this limitation and assumed the big package was a match for this purpose. However I guess that the int64 would work on a 32bit system in this case. As you say there would still be a limit but one that realistically would not be hit. Happy to rework using int64 if you think this is more appropriate...

Patch-set: 4
Reviewer: Gerrit User 25980 <25980@62eb7196-b449-3ce5-99f1-c037f21e1705>