x/crypto/ssh: Fix Protocol Version Exchange (for multi-line)
There is an optional feature of SSH Protocol Version Exchange that
allows metadata lines to be sent prior to the SSH version string.

For example, this is valid:
some-metadata
SSH-2.0-OpenSSH

The current Go implementation takes the first line it sees as
the verison string (in this case, some-metadata). Then, it uses
the next line (SSH-2.0-OpenSSH) as part of key exchange, which
causes sporadic behavior (packet too large) and returns an error.

Unfortunately, this SSH feature is used by some vendors and is part
of the official RFC: https://tools.ietf.org/html/rfc4253#section-4.2

To fix this, I do a simple prefix check each time '\n' is hit. If the
line does not start with SSH-, then I re-allocate it and continue. The
outer loop conditional was changed to prevent a busy loop (scanning
lines forever). The RFC says all lines in version exchange cannot
exceed 255 bytes.

For more info, please refer to:
https://github.com/golang/go/issues/23194

Change-Id: I7be61700a07756353875bf43aad09a580ba533ff