Update patch set 8
Patch Set 8:

> > By redirecting to HTTPS you mean 302 while preserving path and the query, but only for HEAD/GET requests and some client error for POST/PUT/etc?
> 
> Yeah, I think an error (400?) for anything that's not GET/HEAD makes sense. It could also be a 307, but that "fails open" to plaintext instead of immediately making it clear to the client that something is broken and they need to switch to HTTPS for whatever they are POSTing.

Sorry to butt in, but the same can be said for 302+400: if your app is an API endpoint and you're sending your API token via header or URL path, you're also in the same scenario of silent information leakage. I.e.: if this is a concern, then it probably applies to any type of "automatic" redirect.

Patch-set: 8
CC: Gerrit User 26201 <26201@62eb7196-b449-3ce5-99f1-c037f21e1705>